batched_simplerend: Pass derivs to assign_all_from
This completely pointless allocation is obviously not a good solution, but it seems to be required by the current API of assign_all_from_scalar, which assumes the derivs are present in the source when has_derivs() is true.

A better solution would be a version of assign_all_from_scalar that doesn't require the derivatives ... either allowing the derivatives to be set separately, or just assuming they are zero.  Actually, when would you ever want assign_all_from_scalar to not zero the derivatives?  If you want to set all the lanes to the same value, that strongly implies that the value isn't varying, in which case the derivative will be zero.

Signed-off-by: Daniel Dresser <danield@image-engine.com>