ci: Fix broken Mac CI
Make sure LLVM_ROOT/bin (either as a cmake var or as an env var) end
up i the hints list wen searching for llvm-config.

Without this, one of our Mac CI jobs was ending up with clang 14 and
llvm 15, and that was causing problems. The particular test was
supposed to be using 14 for both.

Signed-off-by: Larry Gritz <lg@larrygritz.com>